Move CDiskBlockPos from chain to flatfile.
Cherry-picked from: d6d8a78f26f52fdfe43293686135e2fc6919926c
